api/requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
crypto==1.4.1
cryptography==2.4.2
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysql==0.0.2
mysqlclient==1.3.13
Naked==0.1.31
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pycrypto==2.6.1
pylint==2.1.1
PyMySQL==0.9.2
python-ldap==3.1.0
PyYAML==3.13
requests==2.20.1
shellescape==3.4.1
six==1.11.0
SQLAlchemy==1.2.14
tornado==5.1.1
urllib3==1.24.1
virtualenv==16.1.0
wrapt==1.10.11
